+++ jgclark.Dashboard/CHANGELOG.md
@@ -10,7 +10,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 -->
 
 ## [2.4.0.b20] 2026-02-??? 
-- fix clicking on note title in Task Dialog box 
+- fix clicking on note title in Task Dialog box opening two notes
 - allow Searches to work against either V2 or V3 of Search Extensions plugin
 - move "Default Dashboard Window Type" setting to main Dashboard settings dialog, from the NP plugin settings pane.
 - new setting "Show only projects with next actions" for the Active Projects section. (This includes any where the Project is defined as being 'Sequential'.)